--- a/core/trino-spi/src/main/java/io/trino/spi/block/LazyBlock.java
+++ b/core/trino-spi/src/main/java/io/trino/spi/block/LazyBlock.java
@@ -254,6 +254,12 @@ public class LazyBlock
     }
 
     @Override
+    public boolean mayHaveNull()
+    {
+        return getBlock().mayHaveNull();
+    }
+
+    @Override
     public final List<Block> getChildren()
     {
         return singletonList(getBlock());