Comparable changes to https://github.com/prestodb/presto/pull/16459

Consults `Block#mayHaveNull()` outside of the main processing loop in `HashSemiJoinOperator` so that `Block#isNull` checks can be skipped inside the loop when none are necessary.

Also adds:
- `LazyBlock#mayHaveNull()` implementation to make the `Block#mayHaveNull()` checks like this one more effective
- `Page#getLoadedPage(int... columns)` helper methods to support simultaneous column extraction and `LazyBlock` unwrapping