@raunaqmorarka 2021-07-22
Could you please add JMH benchmark results before and after the change showing the improvement ?