@pettyjamesm 2021-07-22
> Could you please add JMH benchmark results before and after the change showing the improvement ?

It's going to take some time to put that together since no JMH benchmark exists for `HashSemiJoinOperator`. I tried using `SqlSemiJoinInPredicateBenchmark` but unfortunately that query is converted to an inner-join by the planner, so I modified that benchmark to use the query string: `"SELECT orderkey FROM lineitem WHERE orderkey < 0 OR orderkey IN (SELECT orderkey FROM orders WHERE orderkey % 2 = 0)"` and ran it locally with 100 warm-up iterations and 25 measured ones. Run to run variability is high and this benchmark is much less scientific than a JMH based approach would be, but the output looks something like this:

Before:
```
peak_memory:1410102,elapsed_millis:102,input_rows_per_second:735943,output_rows_per_second:294181,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:102147725,cpu_nanos:99034000,user_nanos:98101000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
peak_memory:1410102,elapsed_millis:127,input_rows_per_second:591470,output_rows_per_second:236430,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:127098501,cpu_nanos:109192000,user_nanos:107764000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
peak_memory:1410102,elapsed_millis:139,input_rows_per_second:538891,output_rows_per_second:215413,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:139499266,cpu_nanos:106468000,user_nanos:105621000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
peak_memory:1410102,elapsed_millis:95,input_rows_per_second:789199,output_rows_per_second:315469,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:95254783,cpu_nanos:92582000,user_nanos:91866000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
peak_memory:1410102,elapsed_millis:103,input_rows_per_second:730367,output_rows_per_second:291952,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:102927612,cpu_nanos:97680000,user_nanos:96275000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
peak_memory:1410102,elapsed_millis:114,input_rows_per_second:657643,output_rows_per_second:262882,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:114309571,cpu_nanos:106110000,user_nanos:104536000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
                    sql_semijoin_in ::  106.720 cpu ms :: 1.34MB peak memory :: in 75.2K,      0B,    704K/s,      0B/s :: out 30.1K,   264KB,    282K/s,  2.42MB/s
```

After:
```
peak_memory:1410102,elapsed_millis:78,input_rows_per_second:963263,output_rows_per_second:385049,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:78041971,cpu_nanos:77661000,user_nanos:77446000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
peak_memory:1410102,elapsed_millis:77,input_rows_per_second:970159,output_rows_per_second:387805,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:77487229,cpu_nanos:77326000,user_nanos:77172000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
peak_memory:1410102,elapsed_millis:86,input_rows_per_second:878736,output_rows_per_second:351260,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:85548977,cpu_nanos:84220000,user_nanos:83745000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
peak_memory:1410102,elapsed_millis:87,input_rows_per_second:864962,output_rows_per_second:345754,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:86911267,cpu_nanos:86610000,user_nanos:86282000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
peak_memory:1410102,elapsed_millis:81,input_rows_per_second:925078,output_rows_per_second:369785,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:81263382,cpu_nanos:81105000,user_nanos:80900000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
peak_memory:1410102,elapsed_millis:82,input_rows_per_second:911462,output_rows_per_second:364342,input_megabytes:0,input_megabytes_per_second:0,wall_nanos:82477310,cpu_nanos:82269000,user_nanos:81981000,input_rows:75175,input_bytes:0,output_rows:30050,output_bytes:270450
                    sql_semijoin_in ::   82.769 cpu ms :: 1.34MB peak memory :: in 75.2K,      0B,    908K/s,      0B/s :: out 30.1K,   264KB,    363K/s,  3.12MB/s
```

Logically, these changes here are extending the pattern of using `mayHaveNull()` to unswitch primary loop `isNull` checks like was done in https://github.com/trinodb/trino/pull/8437