@raunaqmorarka 2021-07-22
I think it would be nice to add a JMH benchmark if one doesn't exist while making such optimizations. The linked PR doesn't have JMH numbers either.
Maybe we could go one step further and create 2 separate even more streamlined for loops for the never-null and may-have-null cases. Having a JMH benchmark would help to guide to what extent we should make optimizations like that.