@pettyjamesm 2021-07-22
> Maybe we could go one step further and create 2 separate even more streamlined for loops for the never-null and may-have-null cases. Having a JMH benchmark would help to guide to what extent we should make optimizations like that.

I chose to leave the loop unswitching decision up to the C2 JIT instead of manually unswitching the loops here because a  manually unswitched version would probably want variants for `#mayHaveNull()` _and_ `probeHashChannel >= 0` (and maybe `channelSet.hasNull()` and `channelSet.isEmpty()` too?) which would be at least 4 different loop variants. Having done that exercise, I can say the even with just 2 different patterns, the resulting code is _very_ difficult to read and reason about. While so while C2 may be more fickle about unswitching on both loop invariants than a manual approach, it seemed like the right compromise in this context.

> I think it would be nice to add a JMH benchmark if one doesn't exist while making such optimizations. The linked PR doesn't have JMH numbers either.

I agree, but on my end its been a question of available time and the effort required to produce meaningful benchmarks. In the case of the `PartitionedOutputOperator` PR, I chose not to attach numbers for the existing JMH benchmarks would have required a _lot_ of changes to better simulate real-world conditions and without that work the existing benchmark numbers would be misleading at best. For example, the existing benchmarks don't exercise null channel or any-row replication and don't vary block or `PartitionFunction` implementation classes, which allows the JIT to optimistically inline things deeply in ways that a real worker JVM instance would not.