@pettyjamesm 2021-07-27
Renamed `getLoadedColumns` to `getLoadedPage` and added an inline comment about why `channelSet` was added as a local variable instead of referring to the instance field directly based on review feedback on the PrestoDB PR. The changes should be ready for final review / merge.